requirements.txt
tensorflow>=2.10
fastai>=2.7
numpy>=1.23
matplotlib>=3.5
seaborn>=0.12
Pillow>=9.0
kaggle
jupyter
ipykernel
